Scout: ask_play_cricket synthesises; Reports tab reuses chat report card
ask_play_cricket: drop the calls[] dump. The sub-agent now SYNTHESISES the
gathered data into the exact answer the caller asked for, returning just
{ answer: string }. Earlier shape returned every raw pc_* tool result —
which dumped 80+ fixture rows into the main agent for a question like "find
the one fixture between us and Morpeth on 9 May". The new contract pushes
the filtering into the sub-agent (where it belongs) and treats the
sub-agent's final reply as the authoritative answer. Prompt rewritten to
match: "your reply IS the answer; bullet/table form, IDs included, no JSON
dumps, no narrative".

Tool description on the main-agent side updated so the agent knows to ASK
SPECIFICALLY for what it needs — "find the fixture between us and X on Y"
gets one row, "list all our 1st XI fixtures this season" gets the full
list.

Report card: extracted the inline pipeline / ready / failed cards out of
message-view.tsx into report-card.tsx. ReportsView now renders the same
ReportCard per row (with the polling hook driving live state for in-flight
rows), with a thin footer for the created-at timestamp, thread link, and
delete button. Same visual style across both surfaces; in-flight reports
show a live pipeline + Stop button on the Reports tab too.
