+++ apps/api/src/features/scout/tools/ask-play-cricket.ts
@@ -14,38 +14,16 @@ import { createPlayCricketTools } from "./play-cricket.ts";
  * Same rationale as ask_db (see tools/ask-db.ts): wrap a multi-step external
  * fetcher in its own short-context sub-agent so failed tool calls, dense
  * field-path projections, and the site_id/club_id chain never reach the main
- * chat. Specifically the things the main agent kept tripping on:
+ * chat.
  *
- *   - Projection paths. Each pc_* tool description carries dozens of
- *     dot-notation field paths; over-fetching costs tokens, under-fetching
- *     costs a retry. The sub-agent narrows by trial without polluting chat.
- *   - The status-vs-match_date trap. Played matches stay status="New"
- *     forever; the rule "use match_date vs today" lived in every tool
- *     description and the main agent still got it wrong sometimes.
- *   - The site_id/club_id pivot for opposition scouting (find a Percy Main
- *     vs X match → read X's club_id off it → pass as siteId to
- *     pc_site_matches). Two-call lookup chain that's pure plumbing.
- *
- * Scope vs ask_db: ask_db answers anything our local mirror knows
- * (Percy Main matches past or future, our players, opposition stats in the
- * context of how they've done against us). ask_play_cricket is for
- * everything else — opposition's matches against OTHER clubs, league tables
- * for divisions we're not in, fixtures the local mirror hasn't synced.
- *
- * Output shape: { summary, calls: [{ toolName, output }, ...] }. `calls` is
- * EVERY pc_* data tool result the sub-agent produced, in call order — not
- * just the last one. This matters because PC questions are often plural
- * ("scorecards for these three matches" → three pc_match_detail calls); the
- * earlier "return only the last call" shape ate 2/3 of the answer and made
- * the main agent re-ask question by question. `summary` describes the
- * FETCHING process, not the data — same convention as ask_db.summary. The
- * downstream agent has the cricket context and is responsible for narrative.
+ * Output shape: { answer: string }. The sub-agent runs whatever pc_* calls
+ * it needs internally, then SYNTHESISES the gathered data into the exact
+ * answer the caller asked for. Earlier shape returned every raw pc_* tool
+ * result in `calls[]`; that dumped 50+ fixture rows into the main agent
+ * even when the question was "find one specific match". Synthesis-only
+ * makes the sub-agent commit to its answer instead of pushing the
+ * filtering work upstream.
  */
-
-interface PcCall {
-  toolName: string;
-  output: unknown;
-}
 
 export interface AskPlayCricketToolDeps {
   playCricket: PlayCricketApiClient;
@@ -56,22 +34,12 @@ export interface AskPlayCricketToolDeps {
   logger?: FastifyBaseLogger;
 }
 
-const PC_DATA_TOOL_NAMES = new Set([
-  "pc_list_players",
-  "pc_match_summary",
-  "pc_match_detail",
-  "pc_league_table",
-  "pc_site_matches",
-  "pc_site_results",
-  "pc_find_opposition_matches",
-]);
-
 export function createAskPlayCricketTool(deps: AskPlayCricketToolDeps) {
   const { model } = resolveModel(deps.provider, deps.modelId);
 
   return {
     ask_play_cricket: tool({
-      description: `Answer a question from the Play Cricket public API. Pass a natural-language question; a Play Cricket specialist sub-agent runs the API calls and returns a short \`summary\` plus a \`calls\` array containing the projected payload from EVERY pc_* call it made.
+      description: `Answer a question from the Play Cricket public API. Pass a natural-language question; a Play Cricket specialist sub-agent runs the API calls and returns a single \`answer\` string with the synthesised data the caller asked for.
 
 Use this when the answer needs Play Cricket data NOT in our local DB mirror — typically:
   - Opposition's matches against OTHER clubs (their full season, not just vs us).
@@ -81,14 +49,16 @@ Use this when the answer needs Play Cricket data NOT in our local DB mirror — 
 
 Examples (each ONE call to ask_play_cricket, not many):
   - "Find Newcastle CC's last 5 league results in 2025."
-  - "Get scorecards for these three matches: 7262903, 7262908, 7262912 — every batter's name, runs, balls, how out, and player_id." (returns three pc_match_detail entries in \`calls\`)
+  - "Get scorecards for these three matches: 7262903, 7262908, 7262912 — every batter's name, runs, balls, how out, and player_id."
   - "Who's on Backworth CC's roster — names and play_cricket_ids."
 
-BATCH related questions into one call. The sub-agent will happily make several pc_* calls and return them all in \`calls\` — that's exactly what it's for. Calling ask_play_cricket once per match wastes the sub-agent's batching ability and runs up its setup cost N times.
+ASK SPECIFICALLY for what you need — the sub-agent only returns what the question explicitly asks for. "Find the fixture between us and Morpeth on 9 May" gets ONE row back; "list all our 1st XI fixtures this season" gets the full list. Don't expect raw API dumps — if you need a specific projection (every batter's name, runs, how_out), name those fields in the question.
+
+BATCH related questions into one call where it makes sense — "scorecards for these three matches" works, asking three times wastes the sub-agent's setup cost.
 
 DEFAULT TO ask_db FIRST. The local DB mirrors Play Cricket data for matches Percy Main has played in. Only call ask_play_cricket when ask_db cannot answer — i.e. you need data that doesn't involve a Percy Main fixture, or ask_db's summary said the data isn't there yet.
 
-The sub-agent has no chat context — be specific (club name, season, team/XI, format). \`calls\` is an array of \`{ toolName, output }\` entries, in call order. matchIds and playerIds you'll pass to cite_match / cite_player_stats are in those outputs (sub-agent is required to project them). \`summary\` is metadata about the FETCHING PROCESS (filters applied, the site_id path taken) — read it to verify assumptions and to spot caveats like "player_id only present in roster, not innings.bat[]", but do NOT parrot it back to the user.`,
+The sub-agent has no chat context — be specific (club name, season, team/XI, format). The \`answer\` it returns IS the data — IDs, dates, names, numbers, in compact structured form. matchIds and playerIds you'll pass to cite_match / cite_player_stats are inside the answer.`,
       inputSchema: z.object({
         question: z
           .string()
@@ -113,9 +83,7 @@ The sub-agent has no chat context — be specific (club name, season, team/XI, f
           year: "numeric",
         });
 
-        const system = `You are a Play Cricket API fetcher for Percy Main CSC's Scout system. You call the pc_* tools and return raw projected data. A separate downstream agent reads your data and writes the prose reply with cricket context — you do NOT write that prose yourself. Your final reply text is consumed as a short metadata note about the FETCHING PROCESS, not as a narrative of the data.
-
-The downstream agent receives EVERY pc_* call you make, in order — not just the last one. So if a question asks for N matches' scorecards, call pc_match_detail N times and the downstream agent gets all N. Don't try to compress multiple matches into one call.
+        const system = `You are a Play Cricket API fetcher for Percy Main CSC's Scout system. You call the pc_* tools to gather data, then SYNTHESISE it into the exact answer the caller asked for. The caller does NOT see your raw tool outputs — only your final reply. Your final reply IS the answer.
 
 Today is ${iso} (${ddmmyyyy} dd/mm/yyyy). Current season is ${today.getFullYear()}. The Play Cricket API emits match_date in dd/mm/yyyy — pass values through unchanged.
 
@@ -132,29 +100,28 @@ Each tool's description carries the available field-path projections — read th
 
 CRITICAL rules — apply on every call:
 
-1. **ALWAYS include identifying IDs in the projection.** matches[].id / match_details[].id / players[].player_id (and home_club_id / away_club_id when relevant). The downstream agent uses these IDs to cite sources via cite_match / cite_player_stats — without them no citation can be rendered.
+1. **ALWAYS include identifying IDs in your projections.** matches[].id / match_details[].id / players[].player_id (and home_club_id / away_club_id when relevant). The caller needs these to cite sources downstream — your synthesised answer MUST carry the IDs through.
 
 2. **status is unreliable for played-vs-not-played.** Many played matches retain status "New" indefinitely. To decide whether a match was played, use match_date vs today (${ddmmyyyy}). To confirm a result actually exists, prefer pc_site_results or check that innings are populated in pc_match_detail.
 
 3. **site_id == club_id.** To scout opposition X: find a Percy Main vs X match via pc_match_summary, read X's club_id off the row (home_club_id or away_club_id, whichever isn't 134), pass it as siteId to pc_site_matches / pc_site_results.
 
-4. **Stop when the data is right.** After the final pc_* call has the answer payload, write your short metadata reply and STOP — don't keep poking. You have ${deps.maxSteps} steps total.
+4. **Stop when the data is right.** After the final pc_* call has the answer payload, synthesise your reply and STOP — don't keep poking. You have ${deps.maxSteps} steps total.
 
-OUTPUT — what to write as your final reply (one short sentence, sometimes two):
+OUTPUT — your final reply IS the answer the caller will use. Write it as compact structured data:
 
-DO write:
-  - filters and assumptions: "Filtered to game_type=League and season=2025; took the 5 most-recent fixtures."
-  - schema choices: "Used pc_site_results (cheaper than per-match detail) — only innings totals were needed."
-  - the lookup path for opposition: "Found Newcastle's club_id=152 from Percy Main's 2nd XI fixture on 14/06/2025; pulled their season fixtures from there."
-  - failure path: "No Backworth fixtures returned for 2025; their site_id may have changed." or "API returned an empty matches array."
+DO:
+  - Return ONLY the rows / fields / values the question asked for. If the question asks for one specific match, return one match — even if pc_match_summary returned 80 fixtures along the way.
+  - Include the identifying IDs you projected (match.id, player_id, etc.) — the caller cites against them.
+  - Use compact form: bullet list, table, or one-line-per-record. Plain text or markdown. NO JSON dumps, NO API-internal field paths.
+  - Short note after the data on filters / lookups / caveats: "Filtered season=2026 game_type=League." or "ground_name empty in API — coords unavailable."
 
-DO NOT write:
-  - the data itself or any narrative of it (the downstream agent formats and editorialises)
-  - markdown tables of matches or players
-  - restating the question
-  - empty filler ("Here are the results.")
+DO NOT:
+  - Dump every pc_* tool output verbatim. The caller does not need 80 fixtures when they asked for one.
+  - Editorialise or write narrative cricket prose ("a strong opening partnership", etc.) — leave that to the caller.
+  - Restate the question or write empty filler ("Here are the results:").
 
-If the data IS the entire answer (e.g. one league table), just say what filter/lookup you used.`;
+If the question is genuinely unanswerable (no data, malformed input, ambiguous), say so in one sentence — no fabrication.`;
 
         const startedAt = Date.now();
         let result;
@@ -177,25 +144,21 @@ If the data IS the entire answer (e.g. one league table), just say what filter/l
           };
         }
 
-        // Collect EVERY pc_* tool result in call order. Earlier shape
-        // returned only the last call, which silently dropped 2/3 of the
-        // answer when a question covered multiple matches. Done inline
-        // because generateText's `steps` is typed against the specific
-        // ToolSet passed in, and a helper would either need to be generic
-        // or use an unsafe cast.
-        const calls: PcCall[] = [];
+        const elapsedMs = Date.now() - startedAt;
+        const stepCount = result.steps.length;
+        const answer = result.text.trim();
+
+        // Diagnostic only — count which pc_* tools fired so we can spot a
+        // sub-agent that's over- or under-fetching. The data itself is not
+        // returned; the synthesised `answer` is the only output.
+        const pcCallNames: string[] = [];
         for (const step of result.steps) {
           for (const part of step.content) {
-            if (
-              part.type === "tool-result" &&
-              PC_DATA_TOOL_NAMES.has(part.toolName)
-            ) {
-              calls.push({ toolName: part.toolName, output: part.output });
+            if (part.type === "tool-result") {
+              pcCallNames.push(part.toolName);
             }
           }
         }
-        const elapsedMs = Date.now() - startedAt;
-        const stepCount = result.steps.length;
 
         deps.logger?.info(
           {
@@ -207,24 +170,20 @@ If the data IS the entire answer (e.g. one league table), just say what filter/l
             elapsedMs,
             inputTokens: result.usage.inputTokens ?? 0,
             outputTokens: result.usage.outputTokens ?? 0,
-            callCount: calls.length,
-            toolNames: calls.map((c) => c.toolName),
+            answerChars: answer.length,
+            pcCallNames,
           },
           "scout: ask_play_cricket sub-agent done",
         );
 
-        const summary = result.text.trim();
-        if (!summary && calls.length === 0) {
+        if (!answer) {
           return {
             error:
               "Play Cricket query helper returned no answer. The question may be ambiguous; try rephrasing.",
           };
         }
 
-        return {
-          summary: summary || null,
-          calls,
-        };
+        return { answer };
       },
     }),
   };
